Move Message class into new threads module incl. tests
