+++ tests/test_core.py
@@ -1,4 +1,3 @@
-import asyncio
 import dbm
 import shutil
 import sys
@@ -9,41 +8,8 @@ from pathlib import Path
 from unittest.mock import Mock
 
 import aiodbm
-from aiodbm.core import _Message
 
 python_version = f"{sys.version_info.major}{sys.version_info.minor}"
-
-
-class TestMessage(unittest.IsolatedAsyncioTestCase):
-    async def test_str(self):
-        def alpha():
-            pass
-
-        # given
-        future = asyncio.get_running_loop().create_future()
-        message = _Message(future, alpha)
-        # then
-        self.assertIn("alpha", str(message))
-
-    def test_can_create_stop_signal(self):
-        # when
-        message = _Message.create_stop_signal()
-        # then
-        self.assertTrue(message.is_stop_signal)
-
-    def test_future_strict_raises_error_when_no_future(self):
-        # given
-        message = _Message.create_stop_signal()
-        # when/then
-        with self.assertRaises(ValueError):
-            message.future_strict
-
-    def test_func_strict_raises_error_when_no_func(self):
-        # given
-        message = _Message.create_stop_signal()
-        # when/then
-        with self.assertRaises(ValueError):
-            message.func_strict
 
 
 class DbmAsyncioTestCase(unittest.IsolatedAsyncioTestCase):
